Switch `amount` type from `Double` to `BigDecimal` across models, services, and migrations
- Update models (`Invoice`, `Reimbursement`, `InvoiceCreate`, etc.) to use `BigDecimal` for `amount` fields.
- Adjust related logic in `InvoiceServiceLive`, including reimbursement calculations for precision.
- Modify schema migrations to alter `amount` type in `InvoiceGel`.
- Expand test cases to include scenarios with non-integer and precise `amount` values.

+++ domain/src/scala/domain/models/invoice/Invoice.scala
@@ -8,7 +8,7 @@ import scala.jdk.CollectionConverters.*
 case class Invoice(
   id: InvoiceId,
   name: String,
-  amount: Double,
+  amount: BigDecimal,
   date: LocalDate,
   driver: DriverName,
   kind: String,
